+++ src/components/pages/DiagramJson/index.ts
@@ -1,0 +1,1 @@
+export * from './DiagramJson';
