+++ src/components/pages/Inspector/InspectorJSON/index.ts
@@ -1,3 +1,1 @@
-import { InspectorJSON } from './InspectorJSON';
-
-export default InspectorJSON;
+export * from './InspectorJSON';
