+++ src/components/pages/Inspector/InspectorTable/InspectorTableHeading.tsx
@@ -8,7 +8,7 @@ interface Props {
 const style =
   'px-6 py-3 bg-gray-50 text-left text-xs font-semibold text-malibu-700 uppercase hover:normal-case tracking-wider';
 
-const InspectorTableHeading: FC<Props> = ({
+export const InspectorTableHeading: FC<Props> = ({
   primitiveFeatures,
   headers,
 }) => {
@@ -38,5 +38,3 @@ const InspectorTableHeading: FC<Props> = ({
     </thead>
   );
 };
-
-export default InspectorTableHeading;
