+++ src/components/pages/Inspector/InspectorTable/index.ts
@@ -1,3 +1,1 @@
-import InspectorTable from './InspectorTable';
-
-export default InspectorTable;
+export * from './InspectorTable';
