+++ src/components/pages/Log/index.ts
@@ -1,0 +1,1 @@
+export * from './Log';
