+++ src/components/pages/Tokens/index.ts
@@ -1,0 +1,1 @@
+export * from './Tokens';
