+++ src/components/pages/Workbench/index.ts
@@ -1,0 +1,1 @@
+export * from './Workbench';
